startService ing...

+++ OlympusSDK/Classes/OlympusFileDownloader.swift
@@ -1,7 +1,7 @@
-import Foundation
+
 
 class OlympusFileDownloader {
-    func downloadCSVFile(from url: URL, completion: @escaping (URL?, Error?) -> Void) {
+    func downloadCSVFile(from url: URL, fname: String, completion: @escaping (URL?, Error?) -> Void) {
         let task = URLSession.shared.downloadTask(with: url) { (tempLocalURL, response, error) in
             guard let tempLocalURL = tempLocalURL, error == nil else {
                 completion(nil, error)
@@ -10,9 +10,9 @@ class OlympusFileDownloader {
             
             do {
                 let documentsURL = try FileManager.default.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false)
-                let savedURL = documentsURL.appendingPathComponent("downloadedFile.csv")
+                let savedURL = documentsURL.appendingPathComponent("\(fname).csv")
                 
-                try FileManager.default.removeItem(at: savedURL)
+                if FileManager.default.fileExists(atPath: savedURL.path) { try FileManager.default.removeItem(at: savedURL) }
                 try FileManager.default.moveItem(at: tempLocalURL, to: savedURL)
                 
                 completion(savedURL, nil)
